chore: split trading tests into own files, connect wallet once for deal ticket
